rename prototype_views.py to prototype_gallery_views.py
